+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 
+## [0.3.1] - 2015-05-10
+### Fixed
+- missing dependency codeliner/array-reader
+
 ## [0.3.0] - 2015-05-10
 ### Added
 - Support for prooph/proophessor
@@ -33,6 +37,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Runs already in production!
 
 [unreleased]: https://github.com/prooph/link-app-core/compare/v0.3.0...HEAD
+[0.3.1]: https://github.com/prooph/link-app-core/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/prooph/link-app-core/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/prooph/link-app-core/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/prooph/link-app-core/compare/v0.1...v0.1.1